fix: getPageEditUrl util
