+++ packages/spacecat-shared-utils/src/aem.js
@@ -10,8 +10,6 @@
  * governing permissions and limitations under the License.
  */
 
-import { fetch } from './adobe-fetch.js';
-
 /**
  * Delivery types for AEM deployment
  */
@@ -198,92 +196,4 @@ export function detectAEMVersion(htmlSource, headers = {}) {
   ];
   const found = types.find(([, count]) => count === maxMatches);
   return found[0];
-}
-
-/**
- * Determines the AEM CS/AMS page ID for Content API, from the page URL
- * @param {string} pageURL - The URL of the page
- * @param {string} authorURL - The URL of the author instance
- * @param {string} bearerToken - The access token for the author instance
- * @param {boolean} preferContentApi - Whether to prefer the Content API over the PSS, default is
- * false
- * @param {Object} log - The logger object, default is console
- * @return {string|null} - The AEM page ID
- */
-export async function determineAEMCSPageId(
-  pageURL,
-  authorURL,
-  bearerToken,
-  preferContentApi = false,
-  log = console,
-) {
-  log.info(`Fetching HTML from ${pageURL} to retrieve content-page-id for AEM CS Content API mode.`);
-  const htmlResponse = await fetch(pageURL);
-
-  if (!htmlResponse.ok) {
-    return null;
-  }
-
-  const html = await htmlResponse.text();
-
-  // First try to find a content-page-ref meta tag
-  const contentPageRefRegex = /<meta\s+name=['"]content-page-ref['"]\s+content=['"]([^'"]*)['"]\s*\/?>/i;
-  const refMatch = html.match(contentPageRefRegex);
-
-  if (refMatch?.[1]?.trim()) {
-    if (!authorURL || !bearerToken) {
-      // If ref was present but resolution failed, return null per spec
-      log.warn('Content-page-ref found but authorURL or bearerToken is missing, skipping resolution.');
-      return null;
-    }
-    const contentPageRef = refMatch[1].trim();
-    try {
-      const base = preferContentApi
-        ? `${authorURL}/adobe/experimental/aspm-expires-20251231`
-        : `${authorURL}/adobe/experimental/pss`;
-      const resolveUrl = `${base}/pages/resolve?pageRef=${contentPageRef}`;
-      log.info(`Resolving content-page-ref via ${resolveUrl} (preferContentApi=${preferContentApi})`);
-      const resp = await fetch(resolveUrl, {
-        method: 'GET',
-        headers: { Authorization: bearerToken },
-        redirect: 'follow',
-      });
-      if (resp.status === 200) {
-        let pageId = null;
-        if (preferContentApi) {
-          const data = await resp.json();
-          pageId = data?.id || null;
-        } else {
-          const data = await resp.text();
-          pageId = data || null;
-        }
-
-        if (pageId) {
-          log.info(`Resolved pageId: "${pageId}" from JSON directly for ref "${contentPageRef}"`);
-          return pageId;
-        }
-        log.error('resolve response did not contain an "id" property.');
-        return null;
-      } else {
-        log.warn(`Unexpected status ${resp.status} when resolving content-page-ref.`);
-      }
-    } catch (e) {
-      log.error(`Error while resolving content-page-ref: ${e.message}`);
-    }
-    // If ref was present but resolution failed, return null per spec
-    return null;
-  }
-
-  // Fallback to content-page-id meta tag
-  const contentPageIdRegex = /<meta\s+name=['"]content-page-id['"]\s+content=['"]([^'"]*)['"]\s*\/?>/i;
-  const idMatch = html.match(contentPageIdRegex);
-
-  let pageId = null;
-  if (idMatch?.[1]?.trim()) {
-    pageId = idMatch[1].trim();
-    if (pageId) {
-      log.info(`Extracted pageId: "${pageId}" from "content-page-id" meta tag at ${pageURL}`);
-    }
-  }
-  return pageId;
 }
