+++ packages/spacecat-shared-utils/src/index.js
@@ -91,4 +91,6 @@ export {
   getTemporalCondition,
 } from './calendar-week-helper.js';
 
-export { detectAEMVersion, determineAEMCSPageId, DELIVERY_TYPES } from './aem.js';
+export { detectAEMVersion, DELIVERY_TYPES } from './aem.js';
+
+export { determineAEMCSPageId, getPageEditUrl } from './aem-content-api-utils.js';
