+++ packages/spacecat-shared-utils/test/index.test.js
@@ -81,6 +81,7 @@ describe('Index Exports', () => {
     'detectAEMVersion',
     'determineAEMCSPageId',
     'DELIVERY_TYPES',
+    'getPageEditUrl',
   ];
 
   it('exports all expected functions', () => {
